feat: modify relations between entities in sql

+++ src/main/resources/db.changelog.versions/0.0.1/invoices/create_invoices_table.sql
@@ -2,9 +2,9 @@
 --changeset zkollonay:create_invoices_table splitStatements:true endDelimiter:;
 
 create table invoices (
-      id bigint primary key ,
+      id serial primary key ,
       name varchar(255) not null ,
-      balance bigint,
+      balance decimal(19, 4),
       user_id bigint,
       foreign key (user_id) references users(id)
 );
