+++ src/main/resources/db.changelog.versions/0.0.1/transactions/create_transactions_table.sql
@@ -2,8 +2,8 @@
 --changeset zkollonay:create_transactions_table splitStatements:true endDelimiter:;
 
 create table transactions (
-      id bigint primary key ,
-      amount double precision,
+      id serial primary key ,
+      amount decimal(19, 4),
       scheduled_payment boolean default false,
       transaction_date timestamp,
       payment_method_type varchar(255),
